This is a mayor change how connections can be handled:
With this patch, the infrastructure for async communication is implemented and the ASIO TCP is
adapted to be async.

However, be warned: THIS IS UNTESTED CODE!

Also, the Sputnik Inverter (which is currently the only one implemented) is not adapted to
use this, but under-the-hood everything is async, so I expect breakage.

REPEAT: This version will probably not work.

(But as this is 2 weeks of work, I really need to have this in git.)

TODO:
Sending is not yet "async". Planned to be soon.