[Sputnik Inverter] Refactor the command handling -- preparation commit.
The new class CSputnikCommand will handle the commands for the Inverter,
implementing the commands as an kind of strategy pattern, encapsulating the
data handling and the command date.

This commmit adds the class, but does not yet start using it.
(as some prepartion/rework  in other parts of solarpowerlog is needed to fully
support/enable this concept)