[SIGTERM handling] Sigterm-handling now issues broadcast-event
The old implementation had the problem that significant time passed
between SIGTERM and the actual termination of the program.
This commit works in the direction of fixing this:
The signal handler now sends the broadcast event CMD_BRC_SHUTDOWN
to all subscribers of broadcast signals.

Termination has been changed, so that all pending events are parsed
before actual leaving the programm, so CMD_BRC_SHUTDOWN will be
executed.

Missing is the handling in the inverters and datafilters.
(They should clean up as much as possible to avoid races)