[ICommand] introduce fire-and-forget ICommands
Those are ICommands which ICommandTarget is set to NULL. They will
behave as "normal" ICommands but when they should be executed they
will be simply deleted.

This is handy if you e.g use them as callback for an IConnect API
call, but you are not intersted in the result.