[CSharedConnection*] Rewrite of the SharedConnections.
For design description, see the comment CSharedConnection.h

The class is a complete rewrite to get it more sane.

Currently only the atomic-comms-mode (where a inverter hints that several
steps of a communication must take place in that order) is implemented completly,
for the non-atomic-mode (concurrent, the order of the steps is not set)
the receive handling is missing.
(e.g Needed to get the Simluator working with Shared Comms....)

Some debugging done, still experimental.