[CConfigCentral] First version of configuration and config options abstraction
It is cumbersome to keep documentation and code in sync. This class
should couple documentation with the code, and also make the checking of
the configuration and configuration error reporting more straight forward.

The class is designed to be instantiated within the class to be configured,
so that the values can be directly placed into the class' variables (via
pointers)